add scenes
