+++ usagi/images/map3.dat
