+++ usagi/lib/stage1/item3.rb
