+++ usagi/main.rb
@@ -1,11 +1,16 @@
 require 'dxruby'
-require_relative 'lib/director'
+
+require_relative 'scene'
+require_relative 'lib/load_scenes'
+
 
 Window.width = 800  # 32px * 25マス
 Window.height = 600 # 32px * 17マス + 画面下部のスペース(54px)
 
-Window.loop do
-  Director.instance.play
+Scene.set_current_scene(:stage1)
 
+Window.loop do
   break if Input.key_push? K_ESCAPE
+
+  Scene.play_scene
 end
